Extract public serialization helpers
